requirements.txt
google-generativeai
openai
googlesearch-python
pycountry
streamlit